minor: move internal _pformat_dispatch function
